Refactor ModelStub
