+++ src/test/java/seedu/address/logic/commands/contact/CAddCommandTest.java
@@ -6,24 +6,17 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.address.commons.core.GuiSettings;
-import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.ModelStub;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.AddressBook;
-import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.contact.Contact;
-import seedu.address.model.event.Event;
 import seedu.address.testutil.ContactBuilder;
 
 public class CAddCommandTest {
@@ -79,233 +72,6 @@ public class CAddCommandTest {
     }
 
     /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public ReadOnlyAddressBook getInitialAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        // manage versioned addressBook
-        @Override
-        public void commitAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void undoAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void redoAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void clearHistory() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean isUndoable() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean isRedoable() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        // manage contacts
-        @Override
-        public void addContact(Contact contact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasContact(Contact contact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteContact(Contact target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setContact(Contact target, Contact editedContact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void resetContacts() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Contact> getFilteredContactList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredContactList(Predicate<? super Contact> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void bookmarkContactIndexedAt(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void reshuffleContactsInOrder() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void unmarkContactIndexedAt(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        // manage events
-        @Override
-        public void updateContactListByIndex(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addEvent(Event event) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-
-        @Override
-        public boolean hasEvent(Event event) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-
-        @Override
-        public void deleteEvent(Event target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setEvent(Event target, Event editedEvent) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Event> getFilteredEventList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void resetEvents() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredEventList(Predicate<? super Event> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void sortUpcomingFilteredEventList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        public void updateEventListByIndex(Index index) {
-            throw new AssertionError("This method should not be called.");
-        };
-
-        @Override
-        public void linkEventAndContact(Event event, Contact contact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void unlinkEventAndContact(Event event, Contact contact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void unlinkAllContactsFromEvent(Event event) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void rerenderContactCards() {
-            throw new AssertionError("This method should not be called.");
-        }
-        @Override
-        public void rerenderEventCards() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void rerenderAllCards() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void bookmarkEventIndexedAt(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void reshuffleEventsInOrder() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void unmarkEventIndexedAt(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
      * A Model stub that contains a single contact.
      */
     private class ModelStubWithContact extends ModelStub {
@@ -320,24 +86,6 @@ public class CAddCommandTest {
         public boolean hasContact(Contact contact) {
             requireNonNull(contact);
             return this.contact.isSameContact(contact);
-        }
-    }
-
-    /**
-     * A Model stub that contains a single event.
-     */
-    private class ModelStubWithEvent extends ModelStub {
-        private final Event event;
-
-        ModelStubWithEvent(Event event) {
-            requireNonNull(event);
-            this.event = event;
-        }
-
-        @Override
-        public boolean hasEvent(Event event) {
-            requireNonNull(event);
-            return this.event.isSameEvent(event);
         }
     }
