+++ src/test/java/seedu/address/logic/commands/event/EAddCommandTest.java
@@ -6,23 +6,16 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.address.commons.core.GuiSettings;
-import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.ModelStub;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.AddressBook;
-import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.contact.Contact;
 import seedu.address.model.event.Event;
 import seedu.address.testutil.EventBuilder;
 
@@ -88,239 +81,6 @@ class EAddCommandTest {
 
         // different event -> returns false
         assertFalse(addLectureCommand.equals(addExamCommand));
-    }
-
-    /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public ReadOnlyAddressBook getInitialAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-        //settings
-
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-
-        // manage versioned addressBook
-        @Override
-        public void commitAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void undoAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void redoAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void clearHistory() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean isUndoable() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean isRedoable() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-
-        //contacts
-
-        @Override
-        public void addContact(Contact contact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasContact(Contact contact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteContact(Contact target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setContact(Contact target, Contact editedContact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void resetContacts() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Contact> getFilteredContactList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredContactList(Predicate<? super Contact> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void bookmarkContactIndexedAt(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void reshuffleContactsInOrder() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void unmarkContactIndexedAt(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateContactListByIndex(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        // events
-
-        @Override
-        public void addEvent(Event event) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-
-        @Override
-        public boolean hasEvent(Event event) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-
-        @Override
-        public void deleteEvent(Event target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setEvent(Event target, Event editedEvent) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Event> getFilteredEventList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void resetEvents() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredEventList(Predicate<? super Event> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void sortUpcomingFilteredEventList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        public void updateEventListByIndex(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void linkEventAndContact(Event event, Contact contact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void unlinkEventAndContact(Event event, Contact contact) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void unlinkAllContactsFromEvent(Event event) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void rerenderContactCards() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void rerenderEventCards() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void rerenderAllCards() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void bookmarkEventIndexedAt(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
-        @Override
-        public void reshuffleEventsInOrder() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void unmarkEventIndexedAt(Index index) {
-            throw new AssertionError("This method should not be called.");
-        }
     }
 
     /**
